[WiP] unitframes: target: create a scrollframe in the optionpanel and use it as parent for every element
That way we can scroll within the panel and they "get hidden" when not
in the view. Previous attempt with the optionpanel as parent made the
options which were scrolled past "float" outside the panel, which didn't
look good.
For this we made the changes in core and safeguarded some modules
earlies.