core: updates
*change ns.createTitle
titles won't appear on scrollframes, therefore remove the code to place it there. On the other hand add the ability to pass a string and use that for the title.
*change ns.createDescription
automatically set it's place. This removes the need to do it manually later on
*change ns.createColorPicker
use appropiate internal function name.
*add ns.createScrollFrame
Used to create a scrollframe on optionpanels